requirements.txt
# 核心依赖
requests>=2.31.0
feedparser>=6.0.10
pyyaml>=6.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
python-dateutil>=2.8.2

# AI可选依赖(根据需要安装)
# openai>=1.0.0  # 使用OpenAI兼容API时需要
# anthropic>=0.18.0  # 使用Claude API时需要